backend/requirements.txt
fastapi==0.63.0
psycopg2-binary==2.8.6
pydantic==1.8.1
starlette==0.13.6
typing-extensions==3.7.4.3
